Add translations for yes and no.
NOTE: Weirdly enough, the I18n library starts behaving weirdly when
trying to translate yes and no. I had to rename the translation to
'yes_string' to fix missing translation. Furthermore, any translation of no
resulted in the string 'false' in a radio button. I resolved it by making it
an html element.