Remove reports/notification.json route.
We've been experiencing high CPU issues on our 2 dynos and a GET
to notifications.json often comes up in the log files. This isn't to say
that it's responsible for these issues, but it's unecessary. We haven't
used the SMS feature in about 1 year now so I'm deprecating it. The Android
phone in Brazil still hasn't had the code updated so someone will have to
deal with that there...